Update on RAM handling

+++ orchestrator/cots/otb/algorithms/otb_write_images.py
@@ -29,7 +29,7 @@ It defines method mandatory for a processor
 from orchestrator.cots.otb.otb_app_handler import OtbAppHandler
 from orchestrator.common.logger.maja_logging import configure_logger
 from orchestrator.common.maja_exceptions import MajaBusinessException,MajaIOError
-import shutil
+import shutil,os
 LOGGER = configure_logger(__name__)
 
 
@@ -48,15 +48,17 @@ def write_images(img_list, img_filenames):
         else:
             LOGGER.debug("source : " + img_list[f] + " , dest : " + img_filenames[f])
             try:
-                shutil.copyfile(img_list[f], img_filenames[f])
+                if not os.path.exists(img_filenames[f]) or not os.path.samefile(img_list[f], img_filenames[f]):
+                    shutil.copyfile(img_list[f], img_filenames[f])
             except IOError as err:
                 raise MajaIOError(err)
 
+    if len(tmp_img_list_clean) == 0:
+        return
+
     parameters = {"il": tmp_img_list_clean,
-                  "filenames": tmp_filename_list_clean,
-                  "ram": str(OtbAppHandler.ram_to_use),
-                  "nblinesperstrip": 500}
+                  "filenames": tmp_filename_list_clean
+                  }
 
     app = OtbAppHandler("ImageListWriter", parameters)
-
-    return app
+    del app
